+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 **This version only compatible with _Iris_ 1.8.0 above and _Sodium_.**
 
+### 2.1.2
+- It is strongly recommended that you upgrade to this version.
+- Completely solved the rendering issue of Create mod's Schematic and Outline rendering.
+- Add configuration option to replace Create's checkerboard texture for better shader compatibility.
+
 ### 2.1.0
 - Support Create 6.0.9 and Flywheel 1.0.6.
 - Notice: This is not compatible with Flywheel below 1.0.6 due to breaking changes.